+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to Hylo Node (the Hylo server) will be documented in this fi
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Add query for projects (i.e. Post.projectsForContext), for now the only context is a member's profile
+
 ## [1.4.1]
 
 ### Fixed